graphics: implemented image rendering.
so far only tga files are supported, since glfw does provide
support for this file type only.

+++ src/graphics/include/glfw/glfw_graphics_helper.h
@@ -137,7 +137,7 @@ _checkForError()
 }
 
 void
-_computeFps( float &t0, float &t1, float &frames, float &fps )
+_computeFps( float &t0, float &t1, u32 &frames, float &fps )
 {
 	t1 = glfwGetTime();
